move GreenBulletedList component into components/dev-hub/
