+++ src/components/pages/academia/students-educators-details.js
@@ -3,7 +3,7 @@ import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import GreenBulletedList, {
     DefaultBulletText as BulletText,
-} from '../educators/green-bulleted-list';
+} from '~components/dev-hub/green-bulleted-list';
 import AcademiaEducators from '~images/student-spotlight/academia-educators.svg';
 import AcademiaStudents from '~images/student-spotlight/academia-students.svg';
 import Card from '~components/dev-hub/card';
